SimpleBio.Transcript — Method

Transcript(x::String)

Transcription of the imput DNA equence. This function only allow DNA sequence with A, T, C, G bases. If there are other kind of base in the DNA sequence, consider using xTranscript(x::String)

SimpleBio.xTranscript — Method

xTranscript(x::String)

Transcription of the imput DNA equence. This function allow DNA sequences with bases not included in A, T, C, G.

SimpleBio.DNArc — Method

DNArc(x::String)

This is the function for generating a reverse-compliment sequence to the entered sequence in string type.

SimpleBio.RNArc — Method

RNArc(x::String)

This is the function for generating a reverse-compliment sequence to the entered sequence in string type.

SimpleBio.RTanscript — Method

RTanscript(x::String)

Generate the retro transcription of the entered sequence in string type.