Work with nucleotides

SimpleBio.transcript — Function

transcript(x::String)

transcription of the imput DNA equence. This function only allow DNA sequence with A, T, C, G bases. If there are other kind of base in the DNA sequence, consider using xtranscript(x::String)

SimpleBio.xtranscript — Function

xtranscript(x::String)

transcription of the imput DNA equence. This function allow DNA sequences with bases not included in A, T, C, G.

SimpleBio.DNArc — Function

DNArc(x::String)

This is the function for generating a reverse-compliment sequence to the entered sequence in string type.

SimpleBio.RNArc — Function

RNArc(x::String)

This is the function for generating a reverse-compliment sequence to the entered sequence in string type.

SimpleBio.RTranscript — Function

RTranscript(x::String)

Generate the retro transcription of the entered sequence in string type.

SimpleBio.RandSeq — Function

RandSeq(type::String, length::Int64)

Generate random sequence of select type (RNA, DNA or aminoacid). Acceptable parameters of type: DNA: ["DNA", "DNAs", "DNA nucleotides", "DNA chain", "dna", "dnas", "dna nucleotides", "dna chain"] RNA: ["RNA", "RNAs", "RNA nucleotides", "RNA chain", "rna", "rnas", "rna nucleotides", "rna chain"] amino acid: ["AminoAcid", "aminoacid", "amino acid", "peptide", "aa", "AA"]

RandSeq(type::String, length::Int64, num::Int64)

Generate inputted numbers of random sequence of select type (RNA, DNA or aminoacid). Acceptable parameters of type: DNA: ["DNA", "DNAs", "DNA nucleotides", "DNA chain", "dna", "dnas", "dna nucleotides", "dna chain"] RNA: ["RNA", "RNAs", "RNA nucleotides", "RNA chain", "rna", "rnas", "rna nucleotides", "rna chain"] amino acid: ["AminoAcid", "aminoacid", "amino acid", "peptide", "aa", "AA"]

RandSeq(operator::Int64, type::String, length::Int64, num::Int64)

Generate inputted numbers of random sequence of select type (RNA, DNA or aminoacid). operator = 1 to get array returned. Acceptable parameters of type: DNA: ["DNA", "DNAs", "DNA nucleotides", "DNA chain"] RNA: ["RNA", "RNAs", "RNA nucleotides", "RNA chain"] amino acid: ["AminoAcid", "aminoacid", "amino acid", "peptide", "aa", "AA"]

SimpleBio.isdna — Function

isdna(x::String)

Identify is a sequence is DNA sequence in string type.

SimpleBio.isrna — Function

isrna(x::String)

Identify is a sequence is RNA sequence in string type.

SimpleBio.translatedna — Function

translatedna(x::String)

Return the translation result of the input DNA or RNA sequence in 3-letter amino acid codes.

translatedna(operator::Int64, x::String)

Return the translation result of the input DNA or RNA sequence in 1-letter amino acid codes while operator enter =1.