Primer Design

These are useful tools for primer design for PCR!

Calculate melting temperature

SimpleBio.dnatm — Function

dnatm(dnaseq::String)

Calculate the melting temperature of imput DNA sequence. The calculation is assume the annealing occurs under the standard conditions of 50 nM primer, 50 mM Na+, and pH 7.0.